+++ lighthouse-extension/app/src/lightrider-entry.js
@@ -27,7 +27,7 @@ const LR_PRESETS = {
  * @return {Promise<string|Array<string>|void>}
  */
 async function runLighthouseInLR(connection, url, flags, {lrDevice, categoryIDs, logAssets}) {
-  // Certain fixes need to kick-in under LR, see https://github.com/GoogleChrome/lighthouse/issues/5839
+  // Certain fixes need to kick in under LR, see https://github.com/GoogleChrome/lighthouse/issues/5839
   global.isLightRider = true;
 
   // disableStorageReset because it causes render server hang
@@ -69,13 +69,13 @@ async function runLighthouseInLR(connection, url, flags, {lrDevice, categoryIDs,
 }
 
 if (typeof module !== 'undefined' && module.exports) {
-  // Export for importing types into popup.js, require()ing into unit tests.
+  // Export for require()ing into unit tests.
   module.exports = {
     runLighthouseInLR,
   };
 }
 
-// Expose on window for extension, other browser-residing consumers of file.
+// Expose on window for browser-residing consumers of file.
 if (typeof window !== 'undefined') {
   // @ts-ignore
   window.runLighthouseInLR = runLighthouseInLR;
